pimsleur-data/pyproject.toml
[project]
name = "pimsleur-analyzer"
version = "0.1.0"
description = "Analysis scripts for Pimsleur lesson transcripts."
authors = [
    {name = "Simon",email = "simon@simonpeacocks.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "spacy (>=3.8.7,<4.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "spacy-langdetect (>=0.1.2,<0.2.0)",
    "lingua-language-detector (>=2.1.1,<3.0.0)",
    "mistralai (>=0.4.1,<0.5.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "groq (>=0.28.0,<0.29.0)",
    "supabase (>=2.15.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alive-progress (>=3.2.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pimsleur-analyzer"
version = "0.1.0"
description = "Analysis scripts for Pimsleur lesson transcripts."
authors = ["Simon <simon@simonpeacocks.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
spacy = "^3.7.2"
spacy-langdetect = "^0.1.2"
lingua-language-detector = "^2.1.1"
mistralai = "^0.4.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.11.13"
mypy = "^1.10.0"
types-psycopg2 = "^2.9.21.20250516"

